refactor: Calendar 컨테이너 컴포넌트를 별도로 추가
